StatementAutoCloseResultSet move to SqlAgentImpl inner class
